requirements.txt
scapy>=2.4.5
flask>=2.0.0
requests>=2.25.0
psutil>=5.8.0
pandas>=1.3.0
numpy>=1.21.0
matplotlib>=3.4.0
cryptography>=3.4.0
yara-python>=4.0.0
pyyaml>=5.4.0
scipy>=1.7.0
netifaces>=0.11.0
slack-sdk>=3.0.0
twilio>=7.0.0
python-daemon>=2.3.0
flask-socketio>=5.0.0
python-dotenv>=0.19.0
geoip2>=4.5.0
virustotal-python>=0.2.0
pytest>=6.2.0
pytest-cov>=2.12.0
tabulate>=0.8.9
python-whois>=0.7.3
dnspython>=2.2.0